pyproject.toml
[project]
name = "lgeval"
version = "0.1.0"
description = "GraphRAG benchmark harness for lance-graph, Neo4j, and Kuzu"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
neo4j = ["neo4j==6.1.0"]
kuzu = ["kuzu==0.11.3"]
lance_graph = ["lance-graph==0.5.2", "pyarrow==23.0.0"]
bench = ["psutil==7.2.2"]
data = ["pyarrow==23.0.0"]
all = [
  "neo4j==6.1.0",
  "kuzu==0.11.3",
  "lance-graph==0.5.2",
  "psutil==7.2.2",
  "pyarrow==23.0.0",
]

[project.scripts]
lgeval = "lgeval.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
